fix new boolmatrix::reset() routine.
Forgot to set the internal count-of-set-bits to zero when clearing
all the bits.

Scaling toy problem 'w' by a large number seems to help the solver stay
in the vicinity of an initially 100% correct solution.  Maybe there is
a way to set initial values for a scale of w (or x) wrt C1,C2 by looking
at the raw projection {l,u} bounds?  It is almost impossible to guess
these values (for me).

./mcgenx problem I am down to 2-4 false negatives (but it should be
zero).